// web/src/lib/api/generated/models/githubComNgodingvarengMemoriaInternalDeliveryRestDtoUserResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Book API
 * OpenAPI spec version: 1.0
 */

export interface GithubComNgodingvarengMemoriaInternalDeliveryRestDtoUserResponse {
  created_at?: string;
  email?: string;
  email_verified?: boolean;
  id?: string;
  name?: string;
  /** Username is nil until the welcome/onboarding step claims one. */
  username?: string;
}